Split StringReader declarations and definitions
