+++ src/openrct2/libopenrct2.vcxproj
@@ -180,7 +180,7 @@
     <ClInclude Include="core\RTL.h" />
     <ClInclude Include="core\String.hpp" />
     <ClInclude Include="core\StringBuilder.h" />
-    <ClInclude Include="core\StringReader.hpp" />
+    <ClInclude Include="core\StringReader.h" />
     <ClInclude Include="core\Zip.h" />
     <ClInclude Include="Date.h" />
     <ClInclude Include="Diagnostic.h" />
@@ -504,6 +504,7 @@
     <ClCompile Include="core\RTL.ICU.cpp" />
     <ClCompile Include="core\String.cpp" />
     <ClCompile Include="core\StringBuilder.cpp" />
+    <ClCompile Include="core\StringReader.cpp" />
     <ClCompile Include="core\Zip.cpp" />
     <ClCompile Include="core\ZipAndroid.cpp" />
     <ClCompile Include="Date.cpp" />
