+++ src/openrct2/localisation/LanguagePack.cpp
@@ -15,7 +15,7 @@
 #include "../core/RTL.h"
 #include "../core/String.hpp"
 #include "../core/StringBuilder.h"
-#include "../core/StringReader.hpp"
+#include "../core/StringReader.h"
 #include "Language.h"
 #include "Localisation.h"
